move PlaceName into the generic data module
